+++ Changelog.md
@@ -4,6 +4,7 @@
 ### New features
 * Parsing `OleObject` in `GraphicFrame`
 * Parsing `CommentsExtended`
+* Parsing `DocxParagraph#paragraph_id`, `DocxParagraph#text_id`
 
 ## 0.2.0 (2017-03-25)
 ### New features